+++ o8g/Scripts/windows.py
@@ -69,5 +69,7 @@ def showChangelog():
    switchSetting('GameVersion', getGameVersion())
    msg = """What's new in version 0.9.3
 
-- """
+- Cards effects are applied as if they were activated by active the player (Twelve bugfix).
+- Kyosuke and friends' auto abilities are disabled in United Attack.
+- Balanced pre-build decks. """
    askChoice(msg, [], [], ['Close'])