+++ o8g/Scripts/actionLabels.py
@@ -36,7 +36,7 @@ def aclb_Default(args, x = 0, y = 0):
          else:
             return "Attack"
          
-      if not me.isActive and phaseIdx == BlockPhase:
+      if (not me.isActive or tutorial) and phaseIdx == BlockPhase:
          if hasMarker(card, 'Counter-attack'):
             return "Remove from combat"
          else: