+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Card Fighters' Clash Changelog
 
 ## 0.9.3
+Cards effects are applied as if they were activated by active the player (Twelve bugfix).
+Kyosuke and friends' auto abilities are disabled in UA.
+Balanced pre-build decks.
 ## 0.9.2
 Dragging cards now trigger automation actions.
 Automatic phase control option: game will move to the next phase when required.